Since the rates are for nodes and there are no half nodes etc. changed type of Spawn- /LeaveRate to int

+++ ex3broadcast/src/p2p/analysis/AnalysisElements.java
@@ -4,12 +4,12 @@ package p2p.analysis;
  */
 public class AnalysisElements {
 	String time;
-	double spawnRate;
-	double leaveRate;
+	int spawnRate;
+	int leaveRate;
 	int numOfNodes;
 	int dia;
 	
-	public AnalysisElements(String time, double spawnRate, double leaveRate, int numOfNodes, int dia){
+	public AnalysisElements(String time, int spawnRate, int leaveRate, int numOfNodes, int dia){
 		this.time = time;
 		this.spawnRate = spawnRate;
 		this.leaveRate = leaveRate;
